+++ client/Packages/com.beamable/CHANGELOG.md
@@ -7,7 +7,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [?]
+## [Unreleased]
+
+### Fixed
+- Editor `'namespace' but is used like a 'type'` error 
+
 ### Changed
 - Improved baked content performance by keeping data in a single file and limiting number of IO operations.